+++ friends/src/App.jsx
@@ -1,13 +1,13 @@
 import React from 'react';
 import FriendsList from './components/FriendsList';
-import AddFriendForm from './components/AddFriendForm';
+import FriendForm from './components/FriendForm';
 import './App.css';
 import useApi from './hooks/useApi';
 
 const App = props => {
   const [friends, friendsApi] = useApi('http://localhost:5000/friends');
 
-  const handleAddFriend = friendData => {
+  const handleFormSubmit = friendData => {
     friendsApi.create(friendData);
   };
 
@@ -19,7 +19,7 @@ const App = props => {
     <div className="App">
       <h1>Friends</h1>
       <FriendsList data={friends} onRemove={handleRemoveFriend} />
-      <AddFriendForm onSubmit={handleAddFriend} />
+      <FriendForm onSubmit={handleFormSubmit} />
     </div>
   );
 };
